requirements_tests.txt
-r requirements.txt

pytest
pytest-cov
pytest-django
pytest-flakes
pytest-pep8
pytest-sugar
